+++ codex/search/indexes.py
@@ -1,6 +1,6 @@
 """Haystack search indexes."""
 
-from codex._vendor.haystack.fields import CharField, IntegerField
+from codex._vendor.haystack.fields import CharField, FloatField, IntegerField
 from codex._vendor.haystack.indexes import Indexable, ModelSearchIndex
 from codex.models import Comic
 
@@ -40,18 +40,23 @@ class ComicIndex(ModelSearchIndex, Indexable):
     tags = CharField(model_attr="tags__name", null=True)
     teams = CharField(model_attr="teams__name", null=True)
 
+    # This makes issue field searching exact rather than fuzzy
+    issue = CharField(model_attr="issue", null=True)
+
+    # Floats are a better search experience than fixed decimals
+    issue_number = FloatField(model_attr="issue_number", null=True)
+    community_rating = FloatField(model_attr="community_rating", null=True)
+    critical_rating = FloatField(model_attr="critical_rating", null=True)
+
     class Meta:
         """Model & field include list."""
 
         model = Comic
         fields = (
-            "community_rating",
             "created_at",
-            "critical_rating",
             "day",
             "date",
             "decade",
-            "issue_number",
             "issue_suffix",
             "month",
             "monochrome",