+++ codex/views/browser/choices.py
@@ -124,6 +124,8 @@ class BrowserChoicesViewBase(BrowserBaseView):
     def get_object(self):
         """Get the comic subquery use for the choices."""
         search_scores = self.get_search_scores()
+        if search_scores is None:
+            return self.model.objects.none()  # type: ignore
         object_filter = self.get_query_filters(self.model, search_scores, True)
         return self.model.objects.filter(object_filter)  # type: ignore