+++ codex/views/reader/reader.py
@@ -1,5 +1,6 @@
 """Views for reading comic books."""
 
+from comicbox.box import Comicbox
 from django.db.models import F, IntegerField, Value
 from django.urls import reverse
 from drf_spectacular.types import OpenApiTypes
@@ -7,6 +8,8 @@ from drf_spectacular.utils import OpenApiParameter, extend_schema
 from rest_framework.exceptions import NotFound
 from rest_framework.response import Response
 
+from codex.librarian.importer.tasks import LazyImportComicsTask
+from codex.librarian.mp_queue import LIBRARIAN_QUEUE
 from codex.logger.logging import get_logger
 from codex.models import AdminFlag, Bookmark, Comic
 from codex.serializers.reader import ReaderComicsSerializer
@@ -37,24 +40,32 @@ class ReaderView(BookmarkBaseView):
 
     def _get_comics_list(self):
         """Get the reader naviation group filter."""
-        arc_group = self.params.get("arc_group")
+        select_related = ("series", "volume")
+        prefetch_related = ()
 
+        arc_group = self.params.get("arc_group")
         if arc_group == "a":
             # for story arcs
             rel = "story_arc_numbers__story_arc"
             fields = self._COMIC_FIELDS
             arc_name_rel = "story_arc_numbers__story_arc__name"
             arc_pk_rel = "story_arc_numbers__story_arc__pk"
+            prefetch_related = (*prefetch_related, "story_arc_numbers__story_arc")
             arc_index = F("story_arc_numbers__number")
             ordering = ("arc_index", "date", *Comic.ORDERING)
+            arc_pk_select_related = ()
+            arc_pk_only = ("pk",)
         elif arc_group == self.FOLDER_GROUP:
             # folder mode
             rel = "parent_folder"
             fields = (*self._COMIC_FIELDS, "parent_folder")
             arc_pk_rel = "parent_folder__pk"
             arc_name_rel = "parent_folder__name"
+            select_related = (*select_related, "parent_folder")
             arc_index = Value(None, IntegerField())
             ordering = ("path", "pk")
+            arc_pk_select_related = ("parent_folder",)
+            arc_pk_only = arc_pk_select_related
         else:
             # browser mode.
             rel = "series"
@@ -63,18 +74,29 @@ class ReaderView(BookmarkBaseView):
             arc_name_rel = "series__name"
             arc_index = Value(None, IntegerField())
             ordering = Comic.ORDERING
+            arc_pk_select_related = ("series",)
+            arc_pk_only = arc_pk_select_related
 
-        group_acl_filter = self.get_group_acl_filter(Comic)
         arc_pk = self.params.get("arc_pk")
         if not arc_pk:
-            rel += "__comic"
-            arc_pk = self.kwargs.get("pk")
+            # Get the correct arc/folder/series if not submitted with a post.
+            pk = self.kwargs.get("pk")
+            arc_pk_qs = Comic.objects.filter(pk=pk)
+            arc_pk_qs = arc_pk_qs.select_related(*arc_pk_select_related)
+            arc_pk_qs = arc_pk_qs.prefetch_related(*prefetch_related)
+            arc_pk_comic = (
+                arc_pk_qs.annotate(arc_pk=F(arc_pk_rel)).only(*arc_pk_only).first()
+            )
+            arc_pk = arc_pk_comic.arc_pk  # type: ignore
+
+        group_acl_filter = self.get_group_acl_filter(Comic)
         nav_filter = {rel: arc_pk}
 
         qs = (
             Comic.objects.filter(group_acl_filter)
             .filter(**nav_filter)
-            .prefetch_related("story_arc_numbers__story_arc")
+            .select_related(*select_related)
+            .prefetch_related(*prefetch_related)
             .only(*fields)
             .annotate(
                 series_name=F("series__name"),
@@ -198,6 +220,30 @@ class ReaderView(BookmarkBaseView):
         }
         raise NotFound(detail=detail)
 
+    @staticmethod
+    def _lazy_metadata(current, prev_book, next_book):
+        """Get reader metadata from comicbox if it's not in the db."""
+        import_pks = set()
+        if current and not (current.page_count and current.file_type):
+            with Comicbox(current.path) as cb:
+                current.file_type = cb.get_file_type()
+                current.page_count = cb.get_page_count()
+            import_pks.add(current.pk)
+
+        if prev_book and not prev_book.page_count:
+            with Comicbox(prev_book.path) as cb:
+                prev_book.page_count = cb.get_page_count()
+            import_pks.add(prev_book.pk)
+
+        if next_book and not next_book.page_count:
+            with Comicbox(next_book.path) as cb:
+                next_book.page_count = cb.get_page_count()
+            import_pks.add(next_book.pk)
+
+        if import_pks:
+            task = LazyImportComicsTask(frozenset(import_pks))
+            LIBRARIAN_QUEUE.put(task)
+
     def get_object(self):
         """Get the previous and next comics in a group or story arc."""
         # Books
@@ -209,6 +255,8 @@ class ReaderView(BookmarkBaseView):
 
         prev_book = books.get("prev")
         next_book = books.get("next")
+        self._lazy_metadata(current, prev_book, next_book)
+
         books = {
             "current": current,
             "prev_book": prev_book,